Initial Iteration of the Network Character and Finite-State Machine
Network Character has support for changing perspective with M3 and future implementation for attacking with M1, alongside support for RPC.

Finite-State Machine is a Standard C++ State Machine that uses Enums to assign states and requires setup in advance. A preliminary example is available within source. Its initial implementation uses a Timer but this can probably be ignored as it'd need to be rewritten to function online based on how advanced its behaviour is.